remove interference between players

+++ main.js
@@ -45,7 +45,7 @@ function createPiece(type) {
 }
 
 function updateScore() {
-  document.getElementById('score').innerText = tetris.player.score;
+  // document.getElementById('score').innerText = tetris.player.score;
 }
 
 const tetri = [];
@@ -57,24 +57,38 @@ const playerElements = document.querySelectorAll('.player');
   tetri.push(tetris);
 });
 
-document.addEventListener('keydown', event => {
+const keyListener = (event) => {
   [
     [65, 68, 81, 69, 83],
     [72, 75, 89, 73, 74],
   ].forEach((key, index) => {
     const player = tetri[index].player;
-    if (event.keyCode === key[0]) {
-      player.move(-1);
-    } else if (event.keyCode === key[1]) {
-      player.move(1);
-    } else if (event.keyCode === key[2]) {
-      player.rotate(-1);
-    } else if (event.keyCode === key[3]) {
-      player.rotate(1);
-    } else if (event.keyCode === key[4]) {
-      player.drop();
+    if (event.type === 'keydown') {
+      if (event.keyCode === key[0]) {
+        player.move(-1);
+      } else if (event.keyCode === key[1]) {
+        player.move(1);
+      } else if (event.keyCode === key[2]) {
+        player.rotate(-1);
+      } else if (event.keyCode === key[3]) {
+        player.rotate(1);
+      }
+    }
+
+    if (event.keyCode === key[4]) {
+      if (event.type === 'keydown') {
+        if (player.dropInterval !== player.DROP_FAST) {
+          player.drop();
+          player.dropInterval = player.DROP_FAST;
+        }
+      } else {
+        player.dropInterval = player.DROP_SLOW;
+      }
     }
   });
-});
+};
+
+document.addEventListener('keydown', keyListener);
+document.addEventListener('keyup', keyListener);
 
 updateScore();
